+++ MapboxSearch/sdk/src/androidTest/java/com/mapbox/search/CategorySearchEngineIntegrationTest.kt
@@ -189,6 +189,7 @@ internal class CategorySearchEngineIntegrationTest : BaseTest() {
                 )
             ),
             descriptionAddress = "Starbucks, 750 Castro St, Mountain View, California 94041, United States of America",
+            matchingName = "Starbucks",
             center = Point.fromLngLat(-122.08295, 37.38755),
             routablePoints = listOf(
                 RoutablePoint(point = Point.fromLngLat(-122.08295, 37.38755), name = "Address")